requirements.txt
ipykernel
transformers
streamlit
python-dotenv
torch
peft